Add container storage config
Adds overlay and fuse-overlay driver configs (works on fedora 42,
hopefully won't break ubuntu). This addresses the horrible performance
penalty of having to create mapped copies of the base image when
idmapping with the kernel overlay driver as a non-root user.

Startup of mapped containers should now take <1s as opposed to several
minutes.